change: migrate from script to package entry_point
